move freespace endpoint to kolibri.core.device
