+++ MANIFEST.in
@@ -16,7 +16,6 @@ recursive-include kolibri/plugins *
 recursive-include kolibri/test *
 recursive-include kolibri/utils *
 recursive-include kolibri/tasks *
-recursive-include kolibri/localdevice *
 recursive-include kolibri/*/static *.*
 recursive-include kolibri/*/build/ *.json
 recursive-include kolibri/plugins/*/build *.json
