+++ kolibri/core/device/api_urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import include, url
 from rest_framework import routers
 
-from .api import DevicePermissionsViewSet, DeviceProvisionView
+from .api import DevicePermissionsViewSet, DeviceProvisionView, FreeSpaceView
 
 router = routers.SimpleRouter()
 router.register(r'devicepermissions', DevicePermissionsViewSet, base_name='devicepermissions')
@@ -9,4 +9,5 @@ router.register(r'devicepermissions', DevicePermissionsViewSet, base_name='devic
 urlpatterns = [
     url(r'^', include(router.urls)),
     url(r'^deviceprovision/', DeviceProvisionView.as_view({'post': 'create'}), name='deviceprovision'),
+    url(r'^freespace/', FreeSpaceView.as_view({'get': 'list'}), name='freespace'),
 ]
